+++ CHANGELOG.md
@@ -4,6 +4,10 @@ Changelog
 dev
 ---
 
+* Added a new option: `notify_on_all_messages_in_current_buffer`. When set to
+  `on`, you will receive a notification for all messages in the currently
+  active buffer
+  ([#22](https://github.com/s3rvac/weechat-notify-send/issues/22)).
 * Added a new option: `hide_messages_in_buffers_that_match`. It is a
   comma-separated list of regex patterns for names of buffers from which you
   want to receive notifications without messages